requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
chromadb
faiss-cpu
FaissVectorStore
python-dotenv
langchain-groq
langchain-text-splitters
